search frontend: split tokens.ts and decoratedTokens.ts
