+++ client/shared/src/search/query/providers.ts
@@ -2,7 +2,7 @@ import * as Monaco from 'monaco-editor'
 import { Observable, fromEventPattern, of } from 'rxjs'
 import { scanSearchQuery } from './scanner'
 import { map, first, takeUntil, publishReplay, refCount, switchMap, debounceTime, share } from 'rxjs/operators'
-import { getMonacoTokens } from './tokens'
+import { getMonacoTokens } from './decoratedToken'
 import { getDiagnostics } from './diagnostics'
 import { getCompletionItems } from './completion'
 import { getHoverResult } from './hover'
